+++ exolve-from-text.js
@@ -24,7 +24,7 @@ SOFTWARE.
 The latest code and documentation for Exolve can be found at:
 https://github.com/viresh-ratnakar/exolve
 
-Version: Exolve v1.58 July 10, 2024
+Version: Exolve v1.59 December 22, 2024
 */
 
 /**